+++ apps/design-system/src/subjects/views/secrets/secrets.tsx
@@ -1,30 +1,54 @@
 import { useState } from 'react'
 
-import { noop, useTranslationStore } from '@utils/viewUtils'
+import { useTranslationStore } from '@utils/viewUtils'
 
 import { Button, Drawer, Spacer } from '@harnessio/ui/components'
 import {
   CreateSecretFormFields,
   CreateSecretPage,
+  ExistingSecrets,
   SecretCreationType,
+  SecretItem,
+  SecretScope,
   SecretsHeader,
   SecretType
 } from '@harnessio/ui/views'
 
+import { accountSecrets, organizationSecrets } from './secrets-data'
+
 export const SecretsPage = () => {
   const [selectedType, setSelectedType] = useState<SecretType>(SecretType.New)
 
+  // State for existing secrets
+  const [selectedSecret, setSelectedSecret] = useState<SecretItem | null>(null)
+  const [activeScope, setActiveScope] = useState<SecretScope>('account')
+
   const onSubmit = (data: CreateSecretFormFields) => {
-    console.log(data)
+    console.log('Submitted data:', data)
   }
 
-  const renderContent = () => {
+  // Handlers for existing secrets
+  const handleSelectSecret = (secret: SecretItem) => {
+    setSelectedSecret(secret)
+    console.log('Selected secret:', secret)
+  }
+
+  const handleScopeChange = (scope: SecretScope) => {
+    setActiveScope(scope)
+    console.log('Scope changed to:', scope)
+  }
+
+  const handleCancel = () => {
+    console.log('Cancelled')
+  }
+
+  const renderSecretContent = () => {
     switch (selectedType) {
       case SecretType.New:
         return (
           <CreateSecretPage
             onFormSubmit={onSubmit}
-            onFormCancel={noop}
+            onFormCancel={handleCancel}
             useTranslationStore={useTranslationStore}
             isLoading={false}
             apiError={null}
@@ -38,7 +62,17 @@ export const SecretsPage = () => {
           />
         )
       case SecretType.Existing:
-        return <EntityReferenceExample />
+        return (
+          <ExistingSecrets
+            accountSecrets={accountSecrets}
+            organizationSecrets={organizationSecrets}
+            selectedEntity={selectedSecret}
+            activeScope={activeScope}
+            onSelectEntity={handleSelectSecret}
+            onScopeChange={handleScopeChange}
+            onCancel={handleCancel}
+          />
+        )
       default:
         return null
     }
@@ -57,7 +91,7 @@ export const SecretsPage = () => {
 
         <SecretsHeader onChange={setSelectedType} selectedType={selectedType} />
         <Spacer size={5} />
-        {renderContent()}
+        {renderSecretContent()}
       </Drawer.Content>
     </Drawer.Root>
   )
