+++ packages/ui/src/views/entity-reference/index.ts
@@ -1,3 +1,3 @@
-export * from './entity-reference'
-export { default as EntityReference } from './entity-reference'
-export { default as EntityReferenceExample } from './entity-reference-example'
+export * from './entity-reference-component'
+// export * from './entity-reference'
+// export { default as EntityReferenceExample } from './entity-reference-example'
