+++ packages/ui/src/views/secrets/index.ts
@@ -1,3 +1,4 @@
 export * from './create-secret/create-secret'
 export * from './secrets-header'
 export * from './types'
+export * from './existing-secret/existing-secrets'
